@@ -6,7 +6,7 @@ cmake_minimum_required(VERSION 3.21)
 
 # Only set the cxx_standard if it is not set by someone else
 if (NOT DEFINED CMAKE_CXX_STANDARD)
-  set(CMAKE_CXX_STANDARD 98)
+  set(CMAKE_CXX_STANDARD 11)
 endif()
 
 # strongly encouraged to enable this globally to avoid conflicts between
